requirements.txt
Pillow==8.3.2
opencv-python==4.5.3.56
flask
requests
tesseract==0.1.3
pytesseract==0.3.8